--- src/main/java/ru/ezikvice/springotus/homework7/Homework7Application.java
@@ -1,14 +1,0 @@
-package ru.ezikvice.springotus.homework7;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.sql.SQLException;
-
-@SpringBootApplication
-public class Homework7Application {
-
-    public static void main(String[] args) throws SQLException {
-        SpringApplication.run(Homework7Application.class);
-    }
-}
