--- src/main/resources/data.sql
@@ -1,9 +1,0 @@
-insert into genre (name, description) values ('sci-fi', 'Science Fiction');
-insert into genre (name, description) values ('other genre', 'Other Genre');
-insert into genre (name, description) values ('steampunk', 'vaper tales');
-
-insert into book (name, description) values ('good book', 'very good book');
-insert into book (name, description) values ('bad book', 'very bad book');
-
-insert into author(name) values ('Pushkin A. S.');
-insert into author(name) values ('Dostoevskii F. M.');
