+++ src/test/java/ru/ezikvice/springotus/homework8/CommentDaoTests.java
@@ -1,0 +1,39 @@
+package ru.ezikvice.springotus.homework8;
+
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.test.context.junit4.SpringRunner;
+import ru.ezikvice.springotus.homework8.dao.BookRepository;
+import ru.ezikvice.springotus.homework8.dao.BookRepositoryCustom;
+import ru.ezikvice.springotus.homework8.domain.Book;
+import ru.ezikvice.springotus.homework8.domain.Comment;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+@DataMongoTest
+@ComponentScan("ru.ezikvice.springotus.homework8.dao")
+public class CommentDaoTests {
+
+    @Autowired
+    MongoTemplate em;
+
+    @Autowired
+    private BookRepository rep;
+
+    @Autowired
+    private BookRepositoryCustom repCustom;
+
+    @Test
+    public void savingCommentToBook() {
+        Book book = new Book("Very Popular Book", "Book with comments");
+        Comment comment = new Comment("Test Name Author");
+        repCustom.saveComment(book, comment);
+        Assert.assertTrue(book.getComments().contains(comment));
+    }
+}
