+++ src/test/resources/application.yml
@@ -10,26 +10,4 @@ spring:
   shell:
     interactive:
       enabled: false
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:db_test;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false
-    username: sa
-    password: sa
-    initialization-mode: never
-  jpa:
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.H2Dialect
-        format-sql: true
-    hibernate:
-      ddl-auto: create
-    show-sql: true
-
-  # HikariCP settings
-  # spring.datasource.hikari.*
-
-  #60 sec
-    hikari:
-      connection-timeout: 60000
-      maximum-pool-size: 10
